update to latest github bindings
the previous version of the github bindings did not play nicely when using a
proxy (I can illustrate details if desired). the newer version of the bindings
seems to work, with some tweaks to the function names in the bindings.

thanks so much, this lib was super useful!

fwiw, the milestone refs and issue refs do not seem to be rewritten properly,
I am not sure if this could somehow be related to updating these bindings (it
seems not, but idk). specifically, milestones show up like `%4` in comments
and issue refs like `#42` do not remap to the `#42` from gitlab under the new
issue number in github. @ references are remapped properly (yay). I'm sorry
that I did not take the time to fix these, but highlighting for future users
who may be interested in doing so!