replace usernames with links
replace usernames mentioned in issues, pull requests and comments, with
a link pointing back to the original gitlab user. The link text makes it
clear that it points to a gitlab user and not a github user.

This occurs regardless of what's in usermap in the settings. usermap is
now only useful for assigning issues and pull requests to users.