don't check for issues with same name as MRs
previously, a Merge Request would not be migrated if an issue existed
with the same name. This was because Merge Requests could be migrated as
issues, and the goal was to avoid duplicates.
However, there were too many false positives.
Merge Requests are now no longer migrated as issues anyway, so this
check is being removed.